+++ desktop/CHANGELOG.md
@@ -11,6 +11,9 @@
     - In the settings dialog
 - The main nav bar now shows connected tribes for quick navigation
 - In tribe profile view you can now click on a member to view there profile info
+- Updated Pataka onboarding
+  - updated the Pataka dialog to provided users with other options
+  - Users will now be prompted to join a Pātaka upon creating a profile 
 
 
 ### Bug Fixes
@@ -25,6 +28,7 @@
 - fix showing tribes list if no tribes in appbar
 - fix addPersonFormTitle translations
 - Fixed bug causing selected locale to default to English when another language has been selected
+- Updated the popup messages for consistency in style
 
 ## v1.4.2